// src/fin/types.d.ts
export type Margin = {
  market: string;
  limit?: string;
  fin_address: string;
  config: Market;
};

export type Pair = {
  address: string;
  chainID: NETWORK;
  denoms: [Denom, Denom];
  precision: Precision;
  decimalDelta: number;
  margin?: Margin;
  multiswap: boolean;
  pool?: string;
  queue?: string;
  calc?: string;
};
// ...
export type Precision =
  | { significant_figures: number }
  | { decimal_places: number };